Proposal : Move plugins as a Docker pkg
Thanks to the current design, the Plugins system is very general & it
can infact be a pkg that can be reused by other projects if need be.
