+++ pkg/plugins/client.go
