+++ pkg/plugins/client_test.go
