+++ pkg/plugins/discovery.go
