+++ pkg/plugins/discovery_test.go
