+++ volume/plugins/api.go
@@ -1,7 +1,7 @@
 package plugins
 
 import (
-	"github.com/docker/docker/plugins"
+	"github.com/docker/docker/pkg/plugins"
 	"github.com/docker/docker/volume"
 )
